Trie for prefix-freq matching
Trie data structure is made. It does word insertion and collect(node) function finds all the words starting from the provided node.

Trie is a 3-way trie for space efficiency. Must be stored in localStorage which is small (5MiB).

Currently a standalone trie. Must be incorporated with rest of app later.