Fixed an issue with the Trie where typing a single character would match with incorrect prefix.
Created a dropdown menu for a responsive auto suggestion based on the current trie.

Users cannot insert new words into the trie for now. This will be added in the next pr.

Added an enter button next to search bar. This will allow the user to submit words. Hitting the enter key has the same effect.